add TODO to orm

+++ patterns/src/main/java/ru/danilsibgatyllin/system_pattern/orm/User.java
@@ -2,23 +2,23 @@ package ru.danilsibgatyllin.system_pattern.orm;
 
 public class User {
 
-    private int id;
+    private Long id;
 
     private String login;
 
     private String password;
 
-    public User(int id, String login, String password) {
+    public User(Long id, String login, String password) {
         this.id = id;
         this.login = login;
         this.password = password;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
